+++ Emotion_Classification/Emotion_Classification_Body_Tracking_Modality/emotion_classification_body_tracking_modality/inference_functions.py
@@ -1,15 +1,23 @@
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, Union
 
 import numpy as np
-import tensorflow as tf
+import torch
+import pytorch_lightning as pl
+
+from classification_model import SupervisedModel
 
 
 def make_prediction_from_numpy(
-        x,
-        model
+        x: np.ndarray,
+        model: Union[torch.nn.Module, pl.LightningModule],
+        transforms: Optional[Dict[str, Any]] = None,
 ):
     if len(x.shape) <= 1:
-        x = np.expand_dims(x, axis=0)
+        x = np.expand_dims(x, axis=-1)
+    if transforms is not None:
+        x = transforms(x)
 
-    prediction = tf.squeeze(model(x)).numpy()
-    return prediction
+    x = torch.tensor(x).permute(0, 2, 1)
+    prediction = model(x)
+    prediction = prediction.squeeze(0)
+    return prediction.detach().numpy()