+++ Emotion_Classification/Emotion_Classification_Body_Tracking_Modality/emotion_classification_body_tracking_modality/predict.py
@@ -3,17 +3,19 @@ import json
 import logging
 import os
 
-import tensorflow as tf
-
 from conf import (
     CUSTOM_SETTINGS,
+    MODALITY,
     MODALITY_FOLDER,
     EXPERIMENT_ID,
     ID_TO_LABEL,
     REDIS_HOST,
-    REDIS_PORT
+    REDIS_PORT,
+    PUBLISHER_ON
 )
-from emotion_classification_body_tracking_modality.conf import PUBLISHER_ON
+from classifiers.mlp import MLPClassifier
+from utils.init_utils import init_encoder
+from classification_model import SupervisedModel
 from emotionpubsub import init_redis_emocl_pubsub
 
 
@@ -35,17 +37,15 @@ def predict():
     logger.info("User configurations:")
     logger.info(json.dumps(CUSTOM_SETTINGS, indent=4))
 
-    modality = CUSTOM_SETTINGS['dataset_config']['modality'] if (
-            'modality' in CUSTOM_SETTINGS['dataset_config']
-    ) else 'default_modality'
     ckpt_name = (
         f"{EXPERIMENT_ID}_"
         f"{CUSTOM_SETTINGS['dataset_config']['dataset_name']}_"
-        f"{modality}"
+        f"{MODALITY}_"
+        f"{CUSTOM_SETTINGS[MODALITY]['encoder_config']['class_name']}"
     )
 
     if (
-            CUSTOM_SETTINGS["inference_config"]["mode"] == "features" and
+            CUSTOM_SETTINGS[MODALITY]["inference_config"]["mode"] == "features" and
             PUBLISHER_ON
     ):
         raise ValueError("""
@@ -53,17 +53,35 @@ def predict():
                          and running inference processing. Please, use 'end-to-end' mode".
                          """)
 
+    num_classes = CUSTOM_SETTINGS['dataset_config'].get("number_of_labels", 3)
+    if isinstance(num_classes, dict):
+        num_classes = num_classes.get(MODALITY, 3)
+
     # Initialize models:
     # mode == end-to-end: use fine-tuned model from pre-processed data
-    if CUSTOM_SETTINGS["inference_config"]["mode"] == "end-to-end":
+    if CUSTOM_SETTINGS[MODALITY]["inference_config"]["mode"] == "end-to-end":
         # get checkpoint path
-        if "model_path" not in CUSTOM_SETTINGS["inference_config"]:
+        if "model_path" not in CUSTOM_SETTINGS[MODALITY]["inference_config"]:
             supervised_model_checkpoint_path = os.path.join(
                 MODALITY_FOLDER,
                 "supervised_training",
-                f"{ckpt_name}_model.ckpt"
+                f"{ckpt_name}_model"
             )
-        model = tf.keras.models.load_model(supervised_model_checkpoint_path)
+        else:
+            supervised_model_checkpoint_path = CUSTOM_SETTINGS[MODALITY]["inference_config"]["model_path"]
+        # initialize the model
+        encoder = init_encoder(model_cfg=CUSTOM_SETTINGS[MODALITY]["encoder_config"])
+        classifier = MLPClassifier(
+            encoder.out_size,
+            num_classes,
+            hidden=CUSTOM_SETTINGS[MODALITY]['sup_config'].get("dense_neurons", [64]),
+            p_dropout=CUSTOM_SETTINGS[MODALITY]['sup_config'].get("dropout", None)
+        )
+        model = SupervisedModel.load_from_checkpoint(
+            supervised_model_checkpoint_path + ".ckpt",
+            encoder=encoder,
+            classifier=classifier,
+        )
     else:
         raise ValueError("Unexpected inference mode.")
 
@@ -72,14 +90,14 @@ def predict():
         emocl_pubsub = init_redis_emocl_pubsub(
             REDIS_HOST,
             REDIS_PORT,
-            modality,
-            f"{modality}_data_stream",
-            f"{modality}_data",
+            MODALITY,
+            f"{MODALITY}_data_stream",
+            f"{MODALITY}_data",
             model,
             logger,
             ID_TO_LABEL
         )
-        logging.info(f"Listening to {modality}_data_stream channel...")
+        logging.info(f"Listening to {MODALITY}_data_stream channel...")
         emocl_pubsub.start_processing()
     else:
         raise ValueError("Only pub/sub inference is implemented.")