+++ Multimodal_Fusion/Multimodal_Fusion_Layer/multimodal_fusion_layer/conf.py
@@ -98,23 +98,17 @@ if os.path.exists(PATH_CUSTOM_SETTINGS):
     with open(PATH_CUSTOM_SETTINGS, 'r') as f:
         CUSTOM_SETTINGS = json.load(f)
 
-PUBLISHER_ON = config('PUBLISHER_ON', default=CUSTOM_SETTINGS['inference_config'].get('publisher', False), cast=bool)
+PUBLISHER_ON = config('PUBLISHER_ON', default=CUSTOM_SETTINGS['fusion_config'].get('publisher', True), cast=bool)
 
-# DATA_TO_FUSION = CUSTOM_SETTINGS['inference_config'].get('data_to_fusion',
-#                                                          [CUSTOM_SETTINGS["encoder_config"]["input_type"]])
 DATASET = CUSTOM_SETTINGS["dataset_config"]["dataset_name"]
 
 MODALITY = CUSTOM_SETTINGS["dataset_config"].get("modality", "default_modality")
 
+if type(MODALITY) is list and len(MODALITY) > 1:
+    IS_MULTIMODAL = True
+else:
+    IS_MULTIMODAL = False
+
 SUPPORTED_MODALITIES = ['shimmer', 'body-tracking']
 
-OUTPUT_MODALITY_FOLDER = os.path.join(OUTPUTS_FOLDER, DATASET, MODALITY)
-
-
-# CKPT_NAME = (
-#         f"{EXPERIMENT_ID}_"
-#         f"{CUSTOM_SETTINGS['dataset_config']['dataset_name']}_"
-#         f"{MODALITY}_"
-#         f"{CUSTOM_SETTINGS['sup_config']['input_type']}_"
-#         f"{CUSTOM_SETTINGS['encoder_config']['class_name']}"
-#     )
+# OUTPUT_MODALITY_FOLDER = os.path.join(OUTPUTS_FOLDER, DATASET, MODALITY)