+++ stop_multimodal_services.sh
@@ -1,0 +1,17 @@
+#!/bin/bash
+
+# Deleting All dataset, output files
+
+echo "--------------------"
+docker compose down mock-xroom-writer
+echo "--------------------"
+docker compose down inference-data-processing-bm
+echo "--------------------"
+docker compose down inference-data-processing-body-tracking
+echo "--------------------"
+docker compose down fusion-layer
+echo "--------------------"
+sudo rm datasets/test_data/*
+echo "--------------------"
+
+